Updated the writers to bufWriters Moved LogConfig and LogMessage to different File Updated comments
